refactor(capabilities): split the gemini and anthropic caps into identity and runtime modules
Closes #2325.
